Updated garden endpoints

+++ backend/server/mysql/routes/garden.routes.js
@@ -5,26 +5,31 @@ var router = express.Router();
 //create new garden POST Request with x fields
 router.post('/create', (req, res) => {
   let sql = 'INSERT INTO garden(user_id, garden_name, climate) VALUES(?, ?, ?)';
-  let {user_id, garden, climate} = req.body;
+  let sql2 = 'SELECT user_id FROM user where username = ?';
+  let {username, garden, climate} = req.body;
 
-  db.query(sql, [user_id, garden, climate], (err, results) => {
-    if (err) {
-      throw err;
-    }
-    res.send(results);
-    console.log('Garden created...');
+  db.query(sql2, [username], (err, results) => {
+    const user_id = results[0].user_id;
+    db.query(sql, [user_id, garden, climate], (err, results) => {
+      if (err) {
+        throw err;
+      }
+      res.send(results);
+      console.log('Garden created...');
+    });
   });
 });
 
 //delete a garden DELETE Request with garden name
 
 router.delete('/remove', (req, res) => {
-  let sql = 'DELETE FROM garden WHERE garden_name = ?';
+  let sql = 'DELETE FROM garden WHERE garden_name = ? AND user_id IN (SELECT user_id FROM user WHERE username = ?)';
   let garden_name = req.body.garden_name;
+  let username = req.body.username
 
-  db.query(sql, [garden_name], (err, results) => {
+  db.query(sql, [garden_name, username], (err, results) => {
     if (err) {
-      throw err;
+      return res.status(400).json('Error: ' + err);
     }
     res.send(results);
     console.log('Garden deleted...');
@@ -33,13 +38,13 @@ router.delete('/remove', (req, res) => {
 
 //return all gardens for a user
 router.get('', (req, res) => {
-  let sql = 'SELECT * FROM garden JOIN (SELECT user_id FROM user WHERE username = ?) AS users' + 
+  let sql = 'SELECT * FROM garden JOIN (SELECT user_id FROM user WHERE username = ?) AS users ' + 
             'ON garden.user_id = users.user_id';
   let username = req.query.username;
 
   db.query(sql, [username], (err, results) => {
     if (err) {
-      throw err;
+      return res.status(400).json('Error: ' + err);
     }
     res.send(results);
     console.log('Gardens fetched...');
@@ -49,12 +54,12 @@ router.get('', (req, res) => {
 //edit garden name
 router.post('/change-name', (req, res) => {
   let sql = 'UPDATE garden SET garden_name = ?' + 
-            'WHERE user_id IN (SELECT user_id FROM user WHERE username = ?)';
-  let {garden_name, username} = req.body;
+            'WHERE user_id IN (SELECT user_id FROM user WHERE username = ?) AND garden_name = ?';
+  let {garden_name, username, new_name} = req.body;
 
-  db.query(sql, [garden_name, username], (err, results) => {
+  db.query(sql, [new_name, username, garden_name], (err, results) => {
     if (err) {
-      throw err;
+      return res.status(400).json('Error: ' + err);
     }
     res.send(results);
     console.log('Updated garden...');
@@ -67,7 +72,7 @@ router.get('/identify-climate', (req,res) => {
 
   db.query(sql, [req.query.climate], (err, results) => {
     if(err) {
-      throw err;
+      return res.status(400).json('Error: ' + err);
     }
     res.send(results);
     console.log('fetched emails with gardens of climate ' + req.query.climate);
@@ -80,7 +85,7 @@ router.get('/num-gardens', (req,res) => {
   const {username} = req.query;
   
   db.query(sql, username, (err, results) => {
-    if(err) throw err;
+    if(err) return res.status(400).json('Error: ' + err);
      
     res.send(results[0]);
     console.log('fetched number of gardens');
